fix sweepers

+++ sysdig/resource_sysdig_user_test.go
@@ -37,13 +37,13 @@ func init() {
 			ctx := context.Background()
 			user, err := commonClient.GetUserByEmail(ctx, "terraform-test+user@sysdig.com")
 
-			if err != nil {
-				return err
+			if user == nil || err != nil {
+				return nil
 			}
 
 			err = commonClient.DeleteUser(ctx, user.ID)
 			if err != nil {
-				return err
+				return nil
 			}
 			return nil
